requirements.txt
langchain_core==0.3.41
langchain_openai==0.3.7
langgraph==0.3.5
PyGithub==2.6.1
PyGithub==2.6.1
Pygments==2.19.1
python-dotenv==1.0.1

reportlab==4.3.1
